+++ demos/demo-modules/no-events-label/component.ts
@@ -1,0 +1,46 @@
+import {
+  Component,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef
+} from '@angular/core';
+import {
+  CalendarEvent,
+  CalendarViewPeriod,
+  CalendarMonthViewBeforeRenderEvent,
+  CalendarWeekViewBeforeRenderEvent,
+  CalendarDayViewBeforeRenderEvent
+} from 'angular-calendar';
+import { colors } from '../demo-utils/colors';
+
+@Component({
+  selector: 'mwl-demo-component',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  templateUrl: 'template.html'
+})
+export class DemoComponent {
+  view: string = 'month';
+
+  viewDate: Date = new Date();
+
+  events: CalendarEvent[] = [
+    {
+      title: 'Event 1',
+      color: colors.yellow,
+      start: new Date()
+    }
+  ];
+
+  period: CalendarViewPeriod;
+
+  constructor(private cdr: ChangeDetectorRef) {}
+
+  beforeViewRender(
+    event:
+      | CalendarMonthViewBeforeRenderEvent
+      | CalendarWeekViewBeforeRenderEvent
+      | CalendarDayViewBeforeRenderEvent
+  ) {
+    this.period = event.period;
+    this.cdr.detectChanges();
+  }
+}