+++ demos/demo-modules/no-events-label/sources.ts
@@ -1,0 +1,23 @@
+export const sources = [
+  {
+    filename: 'component.ts',
+    contents: {
+      raw: require('!!raw-loader!./component'),
+      highlighted: require('!!raw-loader!highlightjs-loader?lang=typescript!./component')
+    }
+  },
+  {
+    filename: 'template.html',
+    contents: {
+      raw: require('!!raw-loader!./template.html'),
+      highlighted: require('!!raw-loader!highlightjs-loader?lang=xml!./template.html')
+    }
+  },
+  {
+    filename: 'module.ts',
+    contents: {
+      raw: require('!!raw-loader!./module'),
+      highlighted: require('!!raw-loader!highlightjs-loader?lang=typescript!./module')
+    }
+  }
+];