+++ src/modules/common/calendar-date-formatter.provider.ts
@@ -1,17 +1,18 @@
-import { CalendarNativeDateFormatter } from './calendar-native-date-formatter.provider';
+import { CalendarAngularDateFormatter } from './calendar-angular-date-formatter.provider';
 
 /**
- * This class is responsible for all formatting of dates. There are 2 implementations available, the `CalendarNativeDateFormatter` (default) which will use the <a href="https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/Intl" target="_blank">Intl</a> API to format dates, or there is the `CalendarMomentDateFormatter` which uses <a href="http://momentjs.com/" target="_blank">moment</a>.
+ * This class is responsible for all formatting of dates. There are 3 implementations available, the `CalendarAngularDateFormatter` (default) which uses the angular date pipe to format dates, the `CalendarNativeDateFormatter` which will use the <a href="https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/Intl" target="_blank">Intl</a> API to format dates, or there is the `CalendarMomentDateFormatter` which uses <a href="http://momentjs.com/" target="_blank">moment</a>.
  *
  * If you wish, you may override any of the defaults via angulars DI. For example:
  *
  * ```typescript
  * import { CalendarDateFormatter, DateFormatterParams } from 'angular-calendar';
+ * import { DatePipe } from '@angular/common';
  *
  * class CustomDateFormatter extends CalendarDateFormatter {
  *
  *   public monthViewColumnHeader({date, locale}: DateFormatterParams): string {
- *     return new Intl.DateTimeFormat(locale, {weekday: 'short'}).format(date); // use short week days
+ *     return new DatePipe(locale).transform(date, 'EEE', locale); // use short week days
  *   }
  *
  * }
@@ -23,4 +24,4 @@ import { CalendarNativeDateFormatter } from './calendar-native-date-formatter.pr
  * }]
  * ```
  */
-export class CalendarDateFormatter extends CalendarNativeDateFormatter {}
+export class CalendarDateFormatter extends CalendarAngularDateFormatter {}