+++ src/modules/common/calendar-tooltip.directive.ts
@@ -14,7 +14,7 @@ import {
   Renderer2,
   TemplateRef
 } from '@angular/core';
-import { DOCUMENT } from '@angular/platform-browser';
+import { DOCUMENT } from '@angular/common';
 import { Positioning } from 'positioning';
 import { CalendarEvent } from 'calendar-utils';