+++ src/modules/day/calendar-day.module.ts
@@ -8,7 +8,10 @@ import { CalendarDayViewHourSegmentComponent } from './calendar-day-view-hour-se
 import { CalendarDayViewEventComponent } from './calendar-day-view-event.component';
 import { CalendarCommonModule } from '../common/calendar-common.module';
 
-export { CalendarDayViewComponent } from './calendar-day-view.component';
+export {
+  CalendarDayViewComponent,
+  CalendarDayViewBeforeRenderEvent
+} from './calendar-day-view.component';
 
 @NgModule({
   imports: [
@@ -24,6 +27,8 @@ export { CalendarDayViewComponent } from './calendar-day-view.component';
     CalendarDayViewEventComponent
   ],
   exports: [
+    ResizableModule,
+    DragAndDropModule,
     CalendarDayViewComponent,
     CalendarAllDayEventComponent,
     CalendarDayViewHourSegmentComponent,