+++ test/calendar-week-view.component.spec.ts
@@ -180,9 +180,9 @@ describe('calendarWeekView component', () => {
     };
     fixture.componentInstance.events.push(event);
     fixture.componentInstance.refresh.next(true);
-    expect(fixture.componentInstance.eventRows[0].row[0].event).to.deep.equal(
-      event
-    );
+    expect(
+      fixture.componentInstance.view.eventRows[0].row[0].event
+    ).to.deep.equal(event);
     fixture.destroy();
   });
 
@@ -838,51 +838,44 @@ describe('calendarWeekView component', () => {
     expect(stub).to.have.been.calledOnce; // tslint:disable-line
   });
 
-  it('should not called the event times changed callback when the event was not dragged', () => {
+  it('should only call the beforeViewRender output once when refreshing the view', () => {
     const fixture: ComponentFixture<
       CalendarWeekViewComponent
     > = TestBed.createComponent(CalendarWeekViewComponent);
-    fixture.componentInstance.viewDate = new Date('2016-12-08');
-    fixture.componentInstance.events = [
-      {
-        title: 'foo',
-        color: { primary: '', secondary: '' },
-        start: moment('2016-12-08')
-          .add(4, 'hours')
-          .toDate(),
-        end: moment('2016-12-08')
-          .add(6, 'hours')
-          .toDate(),
-        draggable: true
-      }
-    ];
-    fixture.componentInstance.ngOnChanges({ viewDate: {}, events: {} });
-    fixture.detectChanges();
-    document.body.appendChild(fixture.nativeElement);
-    const event: HTMLElement = fixture.nativeElement.querySelector(
-      '.cal-event-container'
+    fixture.componentInstance.refresh = new Subject();
+    fixture.componentInstance.ngOnInit();
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    const beforeViewRenderCalled = sinon.spy();
+    // use subscription to test that it was only called a max of one times
+    const subscription = fixture.componentInstance.beforeViewRender.subscribe(
+      beforeViewRenderCalled
     );
-    const eventPosition: ClientRect = event.getBoundingClientRect();
-    const eventTimesChanged = sinon.spy();
-    fixture.componentInstance.eventTimesChanged.subscribe(() =>
-      eventTimesChanged()
-    );
-    triggerDomEvent('mousedown', event, {
-      clientX: eventPosition.left,
-      clientY: eventPosition.top
-    });
-    fixture.detectChanges();
-    triggerDomEvent('mousemove', document.body, {
-      clientX: eventPosition.left,
-      clientY: eventPosition.top
-    });
-    fixture.detectChanges();
-    triggerDomEvent('mouseup', document.body, {
-      clientX: eventPosition.left,
-      clientY: eventPosition.top
-    });
-    fixture.detectChanges();
+    fixture.componentInstance.refresh.next(true);
+    expect(beforeViewRenderCalled).to.have.callCount(1);
+    subscription.unsubscribe();
     fixture.destroy();
-    expect(eventTimesChanged.callCount).to.equal(0);
+  });
+
+  it('should expose the view period on the beforeViewRender output', () => {
+    const fixture: ComponentFixture<
+      CalendarWeekViewComponent
+    > = TestBed.createComponent(CalendarWeekViewComponent);
+    const beforeViewRenderCalled = sinon.spy();
+    fixture.componentInstance.beforeViewRender
+      .take(1)
+      .subscribe(beforeViewRenderCalled);
+    fixture.componentInstance.ngOnInit();
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.componentInstance.ngOnChanges({ viewDate: {} });
+    expect(
+      beforeViewRenderCalled.getCall(0).args[0].period.start instanceof Date
+    ).to.equal(true);
+    expect(
+      beforeViewRenderCalled.getCall(0).args[0].period.end instanceof Date
+    ).to.equal(true);
+    expect(
+      Array.isArray(beforeViewRenderCalled.getCall(0).args[0].period.events)
+    ).to.equal(true);
+    fixture.destroy();
   });
 });